[move] パッケージを整理する
